Add GenerateOSSpec to generate specs for platforms other than the current
